backend/requirements.txt
# Django core
Django>=5.0,<6.0
djangorestframework>=3.14
django-cors-headers>=4.3
django-filter>=23.5
django-environ>=0.11

# Database
psycopg2-binary>=2.9

# Authentication
djangorestframework-simplejwt>=5.3
drf-spectacular>=0.27

# Utilities
python-dotenv>=1.0
Pillow>=10.0

# Development
black>=24.0
isort>=5.13
flake8>=7.0
pytest>=8.0
pytest-django>=4.8
pytest-cov>=5.0
factory-boy>=3.3

# Production
gunicorn>=21.0
whitenoise>=6.6